+++ src/base/configuration.cpp
@@ -94,10 +94,6 @@ std::string Configuration::copyright() {
   ss << "THIS SOFTWARE IS PROVIDED AS-IS, WITHOUT ANY WARRANTIES.\n"
      << "USE AT YOUR OWN RISK.\n\n";
 
-  ss << "cvc5 incorporates code from ANTLR3 (http://www.antlr.org).\n"
-     << "See licenses/antlr3-LICENSE for copyright and licensing information."
-     << "\n\n";
-
   ss << "This version of cvc5 is linked against the following non-(L)GPL'ed\n"
      << "third party libraries.\n\n";