+++ src/expr/type_checker_template.cpp
@@ -46,7 +46,9 @@ TypeNode TypeChecker::preComputeType(NodeManager* nodeManager, TNode n)
       typeNode = nodeManager->builtinOperatorType();
       break;
 
-      // !!! will auto-generate preComputeType rules when they are available
+      // clang-format off
+${pretyperules}
+      // clang-format on
 
     default:
       // not handled