+++ src/util/result.cpp
@@ -143,7 +143,6 @@ ostream& operator<<(ostream& out, const Result& r) {
   Language language = options::ioutils::getOutputLanguage(out);
   switch (language) {
     case Language::LANG_SYGUS_V2: r.toStreamSmt2(out); break;
-    case Language::LANG_TPTP: r.toStreamTptp(out); break;
     default:
       if (language::isLangSmt2(language))
       {
@@ -182,23 +181,6 @@ void Result::toStreamSmt2(ostream& out) const {
     return;
   }
   toStreamDefault(out);
-}
-
-void Result::toStreamTptp(std::ostream& out) const {
-  out << "% SZS status ";
-  if (d_status == Result::SAT)
-  {
-    out << "Satisfiable";
-  }
-  else if (d_status == Result::UNSAT)
-  {
-    out << "Unsatisfiable";
-  }
-  else
-  {
-    out << "GaveUp";
-  }
-  out << " for " << getInputName();
 }
 
 }  // namespace cvc5::internal